pyproject.toml
[project]
name = "nonebot-plugin-netease-name"
version = "0.1.0"
description = "Nonebot plugin for for generating Minecraft Netease Version nickname"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2>=2.4.2",
]
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "basic"
